- New layout for the frontend with Header, Navigation and a MainContent with a sidebar - removed the user login from the header class, moved it to a component of it's own that is placed into the new navigation area
