+++ frontend/src/components/Header.css
@@ -1,38 +1,6 @@
 .header {
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-    padding: 20px;
-    background-color: #282b35;
-    position: relative;
-    height: 32px;
-}
-
-.header-content {
-    display: flex;
-    align-items: center;
-    gap: 10px;
-}
-
-.header-input {
-    padding: 8px;
-    font-size: 16px;
-}
-
-.header-button {
-    padding: 8px 12px;
-    font-size: 16px;
-    cursor: pointer;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 4px;
-}
-
-.header-button:hover {
-    background-color: #0056b3;
-}
-
-.header-welcome {
-
+    background-color: #212529;
+    color: #ffffff;
+    padding: 10px 20px;
+    text-align: center;
 }
