+++ frontend/src/components/MainContent.css
@@ -1,0 +1,20 @@
+.container {
+    max-width: 1200px;
+    background-color: #343a3f;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+    display: flex;
+    padding: 10px 20px;
+}
+
+.main-content {
+    flex: 1;
+    margin-left: 10px;
+}
+
+.sidebar {
+    width: 300px;
+    background-color: #212529;
+    padding: 20px;
+    margin-right: 10px;
+    box-shadow: 0 0 5px rgba(0, 0, 0, 0.1);
+}
