+++ frontend/src/components/Navigation.css
@@ -1,0 +1,8 @@
+.navigation {
+    display: flex;
+    justify-content: flex-end;
+    align-items: center;
+    background-color: #343a40;
+    color: #ffffff;
+    padding: 10px 20px;
+}
