+++ frontend/src/components/UserLogin.css
@@ -1,0 +1,18 @@
+.login-input {
+    padding: 8px;
+    font-size: 16px;
+    width: 128px;
+}
+
+.login-container {
+    display: flex;
+    justify-content: flex-end;
+    align-items: center;
+    height: 48px;
+    width: 512px;
+    gap: 10px;
+}
+
+.login-button {
+    width: 96px;
+}
